requirements.txt
Flask>=2.0
python-dotenv>=0.21
requests>=2.28
waitress>=2.1
gunicorn>=20.1
PyGithub>=1.58
GitPython>=3.1.30
